Bring updater config into updater package
