+++ ocsp/updater/updater_test.go
@@ -21,7 +21,6 @@ import (
 	bgrpc "github.com/letsencrypt/boulder/grpc"
 	blog "github.com/letsencrypt/boulder/log"
 	"github.com/letsencrypt/boulder/metrics"
-	ocsp_updater_config "github.com/letsencrypt/boulder/ocsp/updater/config"
 	rocsp_config "github.com/letsencrypt/boulder/rocsp/config"
 	"github.com/letsencrypt/boulder/sa"
 	sapb "github.com/letsencrypt/boulder/sa/proto"
@@ -76,7 +75,7 @@ func setup(t *testing.T) (*OCSPUpdater, sapb.StorageAuthorityClient, *db.Wrapped
 		nil,
 		strings.Fields("0 1 2 3 4 5 6 7 8 9 a b c d e f"),
 		&mockOCSP{},
-		ocsp_updater_config.Config{
+		Config{
 			OldOCSPBatchSize:         1,
 			OldOCSPWindow:            cmd.ConfigDuration{Duration: time.Second},
 			SignFailureBackoffFactor: 1.5,
@@ -717,7 +716,7 @@ func mkNewUpdaterWithStrings(t *testing.T, shards []string) (*OCSPUpdater, error
 		nil,
 		shards,
 		&mockOCSP{},
-		ocsp_updater_config.Config{
+		Config{
 			OldOCSPBatchSize:         1,
 			OldOCSPWindow:            cmd.ConfigDuration{Duration: time.Second},
 			SignFailureBackoffFactor: 1.5,
